# docs/openapi.yaml
openapi: 3.0.3
info:
  title: urekoi API
  description: 熟女専門マッチングアプリ urekoi のAPI仕様
  version: 0.1.0
  license:
    name: UNLICENSED
    url: https://github.com/koudai001/urekoi
servers:
  - url: https://github.com/koudai001/urekoi
    description: 本番環境のAPIサーバー。実際のURLは環境変数で切り替える想定。
...
paths:
...
  /myprofile:
    get:
      tags: [myprofile]
      operationId: getMyprofile
      summary: 自分のプロフィール取得
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyProfileResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: プロフィール未作成
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: profile not found
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags: [myprofile]
      operationId: putMyprofile
      summary: プロフィール更新
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MyProfileRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyProfileResponse"
        "400":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: プロフィール未作成
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: profile not found
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    wsTicket:
      type: apiKey
      in: query
      name: ticket
      description: POST /ws/ticketで発行した使い捨てチケット
  responses:
    Unauthorized:
      description: 未認証
    InternalServerError:
      description: サーバー内部エラー
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    MyProfileRequest:
      type: object
      required: [nickname, prefecture_code]
      properties:
        nickname:
          type: string
          example: たろう
        prefecture_code:
          type: integer
          example: 13
        bio:
          type: string
          example: よろしくお願いします
        occupation:
          type: string
          example: エンジニア
        hometown:
          type: string
          example: 東京
        blood_type:
          type: string
          example: A型
        mbti:
          type: string
          example: INTJ
        body_type:
          type: string
          example: 普通
        education:
          type: string
          example: 大学卒
        holiday:
          type: string
          example: 土日
        alcohol:
          type: string
          example: 時々飲む
        smoking:
          type: string
          example: 吸わない
        height_cm:
          type: integer
          example: 175
        tag_ids:
          type: array
          items:
            type: integer
          example: [1, 2, 3]
    MyProfileResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nickname:
          type: string
          example: たろう
        age:
          type: integer
          example: 42
        prefecture_code:
          type: integer
          example: 13
        prefecture:
          type: string
          example: 東京都
        bio:
          type: string
          example: よろしくお願いします
        occupation:
          type: string
          example: エンジニア
        hometown:
          type: string
          example: 東京
        blood_type:
          type: string
          example: A型
        mbti:
          type: string
          example: INTJ
        body_type:
          type: string
          example: 普通
        education:
          type: string
          example: 大学卒
        holiday:
          type: string
          example: 土日
        alcohol:
          type: string
          example: 時々飲む
        smoking:
          type: string
          example: 吸わない
        height_cm:
          type: integer
          example: 175
        tag_ids:
          type: array
          items:
            type: integer
          example: [1, 2, 3]
        images:
          type: array
          description: sort_order順に並んだ登録済みのプロフィール画像
          items:
            $ref: "#/components/schemas/ProfileImageResponse"
...
    ProfileImageResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        url:
          type: string
          example: https://images.urekoi.example.com/profiles/1/xxxx.jpg
        sort_order:
          type: integer
          example: 0
    Error:
      type: object
      properties:
        error:
          type: string